feat: Item 엔티티 averageRating을 Double로 변경 (소수점 지원) feat: Comment 조회 쿼리 최적화

+++ src/main/java/com/teamproject/back/entity/Item.java
@@ -48,7 +48,7 @@ public class Item {
     private List<ItemDetailImage> itemDetailImages = new ArrayList<>();
 
     @Transient // 이 필드는 DB에 저장되지 않음
-    private Integer averageRating;
+    private Double averageRating;
 
     @Transient // 이 필드는 DB에 저장되지 않음
     private Integer commentCount;
